Move importlib.resources to be in 3, but only usable on 3.7
